feat(dto): add LancamentoEstatisticaCategoriaQuantidade and update statistics types

+++ src/main/java/br/com/decodex/gestaofinanceira/dto/estatisticas/LancamentoEstatisticaPessoa.java
@@ -1,5 +1,7 @@
 package br.com.decodex.gestaofinanceira.dto.estatisticas;
 
+import java.math.BigDecimal;
+
 import br.com.decodex.gestaofinanceira.model.Pessoa;
 import br.com.decodex.gestaofinanceira.model.TipoLancamento;
 
@@ -9,9 +11,9 @@ public class LancamentoEstatisticaPessoa {
 
 	private Pessoa pessoa;
 
-	private Long total;
+	private BigDecimal total;
 
-	public LancamentoEstatisticaPessoa(TipoLancamento tipo, Pessoa pessoa, Long total) {
+	public LancamentoEstatisticaPessoa(TipoLancamento tipo, Pessoa pessoa, BigDecimal total) {
 		this.tipo = tipo;
 		this.pessoa = pessoa;
 		this.total = total;
@@ -33,11 +35,11 @@ public class LancamentoEstatisticaPessoa {
 		this.pessoa = pessoa;
 	}
 
-	public Long getTotal() {
+	public BigDecimal getTotal() {
 		return total;
 	}
 
-	public void setTotal(Long total) {
+	public void setTotal(BigDecimal total) {
 		this.total = total;
 	}
 }
